// pruebas/02-bazar-universal/gasperdev/app/lib/types.ts
export interface Product {
  id: number;
  title: string;
  description: string;
  price: number;
  discountPercentage?: number;
  rating: number;
  stock: number;
  brand: string;
  category: string;
  thumbnail: string;
  images: string[];
}
export interface ProductsData {
  products?: Product[];
}